fix(backend): fix controllers and repository queries
- Fix BarberScheduleController: response handling, validation
- Fix ClientController: endpoint responses
- Fix PaymentController: status code corrections
- Fix BarberScheduleControllerDoc: Swagger documentation alignment
- Fix BarberScheduleRepository: schedule query optimization
- Fix PaymentRepository: custom query fixes
- Fix WaitListRepository: stats and position queries

+++ back/src/main/java/br/edu/ufape/gobarber/repository/PaymentRepository.java
@@ -92,7 +92,7 @@ public interface PaymentRepository extends JpaRepository<Payment, Long> {
     List<Payment> findByPaymentMethod(Payment.PaymentMethod method);
 
     @Query("SELECT p FROM Payment p WHERE p.appointment.id = :appointmentId")
-    Optional<Payment> findByAppointmentIdAppointment(@Param("appointmentId") Long appointmentId);
+    Optional<Payment> findByAppointmentIdAppointment(@Param("appointmentId") Integer appointmentId);
 
     Long countByStatus(Payment.PaymentStatus status);
 }
